+++ ChangeLog.md
@@ -10,6 +10,7 @@
     - Fix stylesheets to better manage tabbed nesting
     - Use pwix:logger universal logger, thus bumping minor candidate version number
     - Align border radius on bootstrap measures
+    - Extend link validity to have 'From ... to ...'
 
 ### 1.3.0